Refactoring Total du fichier requetebd5 + ajout de toute les docString de tout les requete dans les classes
