+++ GestLab/initialisation.py
@@ -1,0 +1,6 @@
+from .connexionPythonSQL import *
+
+cnx = ouvrir_connexion()
+
+def get_cnx():
+    return cnx
